+++ text/hibernate6-text/src/main/java/com/github/starnowski/posjsonhelper/text/hibernate6/descriptor/TSVectorFunctionDescriptor.java
@@ -44,7 +44,6 @@ public class TSVectorFunctionDescriptor extends NamedSqmFunctionDescriptor {
 
     @Override
     protected <T> SelfRenderingSqmFunction<T> generateSqmFunctionExpression(List<? extends SqmTypedNode<?>> arguments, ReturnableType<T> impliedResultType, QueryEngine queryEngine) {
-        //TODO Check arguments size
-        return (SelfRenderingSqmFunction<T>) new TSVectorFunction((Path) arguments.get(arguments.size() > 1 ? 1 : 0), arguments.size() > 1 ? (SqmExpression<?>) arguments.get(0) : null, queryEngine.getCriteriaBuilder());
+        return (SelfRenderingSqmFunction<T>) new TSVectorFunction(arguments.get(arguments.size() > 1 ? 1 : 0), arguments.size() > 1 ? (SqmExpression<?>) arguments.get(0) : null, queryEngine.getCriteriaBuilder());
     }
 }